requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
celery==4.3.0
Django==2.2
django-celery-results==1.1.2
django-cors-headers==2.5.2
django-simple-history==2.8.0
django-stubs==1.1.0
djangorestframework==3.9.2
djangorestframework-stubs==0.1.0
importlib-metadata==0.23
jsonpickle==1.1
kombu==4.6.5
more-itertools==7.2.0
mypy==0.730
mypy-extensions==0.4.1
psycopg2==2.8.1
pytz==2018.9
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
vine==1.3.0
zipp==0.6.0
